refactor: change WinningLotto class not to be singleton
