+++ src/main/java/lotto/domain/WinningLotto.java
@@ -1,12 +1,10 @@
 package lotto.domain;
 
-import lotto.domain.Factory.LottoTicketFactory;
+import lotto.domain.factory.LottoTicketFactory;
 import lotto.exception.DuplicatedInputException;
 import lotto.exception.ExceptionMessage;
 
 public class WinningLotto {
-    private static WinningLotto WINNING_LOTTO = null;
-
     private final LottoTicket winningNumbers;
     private final LottoNumber bonusBall;
 
@@ -18,11 +16,7 @@ public class WinningLotto {
     }
 
     public static WinningLotto of(final String winningLotto, final int bonusBall) {
-        if (WINNING_LOTTO == null) {
-            WINNING_LOTTO = new WinningLotto(winningLotto, bonusBall);
-        }
-
-        return WINNING_LOTTO;
+        return new WinningLotto(winningLotto, bonusBall);
     }
 
     public LottoTicket getWinningNumbers() {
@@ -48,9 +42,5 @@ public class WinningLotto {
         if (isDistinct) {
             throw new DuplicatedInputException(ExceptionMessage.DUPLICATED_NUMBER_EXCEPTION);
         }
-    }
-
-    static void makeObjectNull() {
-        WINNING_LOTTO = null;
     }
 }
