+++ src/main/java/lotto/domain/dto/LottoDTO.java
@@ -1,4 +1,4 @@
-package lotto.db.dto;
+package lotto.domain.dto;
 
 import java.util.Arrays;
 import java.util.Collections;
