+++ src/main/java/lotto/domain/factory/LottoTicketFactoryStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain.Factory;
+package lotto.domain.factory;
 
 import lotto.domain.LottoTicket;
