--- src/main/resources/templates/css/custom.css
@@ -1,13 +1,0 @@
-.btn-round {
-	border-radius: 50%;
-	text-align: center;
-	text-decoration: none;
-}
-
-.vertical-center {
-  margin: 0;
-  position: absolute;
-  top: 50%;
-  -ms-transform: translateY(-50%);
-  transform: translateY(-50%);
-}
