--- src/main/resources/templates/previous_lotto.html
@@ -1,39 +1,0 @@
-<!doctype html>
-<html lang="ko">
-<head>
-	<!-- Required meta tags -->
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-	<!-- Bootstrap CSS -->
-	<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-
-	<link rel="stylesheet" type="text/css" href="css/custom.css">
-
-	<title>로또 - 지난 당첨내역</title>
-</head>
-
-<body class="bg-dark">
-	<div id="header"></div>
-
-	<div class="container mt-5">
-		<div class="jumbotron">
-			<h4 class="display-4">지난 당첨내역</h4>
-			<hr class="my-4">
-
-
-			<div class="container" id="purchasing">
-				<a class="btn btn-warning btn-lg btn-block" href="/purchase">로또 구입하기</a>
-				<a class="btn btn-warning btn-lg btn-block" href="/purchase">지난 당첨내역 보기</a>
-				<a class="btn btn-warning btn-lg btn-block" href="/purchase">당첨 통계</a>
-			</div>
-
-		</div>
-	</div>
-
-	<!-- Optional JavaScript -->
-	<!-- jQuery first, then Popper.js, then Bootstrap JS -->
-	<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-	<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-</body></html>
