+++ src/test/java/lotto/domain/factory/LottoTicketFactoryTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.Factory;
+package lotto.domain.factory;
 
 import lotto.domain.LottoTicket;
 import lotto.exception.DuplicatedInputException;
